Tested and fixed many issues with the evaluator strategy.
Code is not pretty at all and needs to be organized. Some efficiency measures can be taken (such as storing values from
previous evaluations). There are three unchecked corner cases that will be added to the issues. Removed binaries. Added
Utils class for fileParsing, will use regex in the future.