Major refactor and cleanup
Quite a lot of refactoring was performed, including package name changes
The documentation is done with the exception of the GUI
Command file parsing is done using regex
The deck file was not being correctly inverted. The order did not match specification. This was fixed.
Added package-info.java files to each package with a brief description.